+++ scintellometry/ephemerides/psrb1957+20.par
@@ -1,33 +1,40 @@
-# received from Ben Stappers
-PSRJ           J1959+2048
-RAJ             19:59:36.7601472         1  0.00007670749855293170   
-DECJ           +20:48:14.90687           1  0.00165532554607184266   
-F0             622.1220287855852541      1  0.00000000004697237789   
-F1             -6.5013752102446370666e-15 1  1.3519178586902677628e-19
-PEPOCH         51260                       
-POSEPOCH       51260                       
-DMEPOCH        51260                       
-DM             29.124889282012632186     1  0.00005807065328697178   
-PMRA           -14.283787399841813343    1  0.09882848300781917061   
-PMDEC          -25.821814586554778983    1  0.15019963242904293965   
+# received from Ben Stappers, 5 Sep 2014
+# [[rmail:/home/mhvk/Post/scintellometry.xmail#1BFB8327-2DC9-4F37-B9C1-F369E8CA9608@manchester.ac.uk][Email from Benjamin Stappers: Re: 1957+20 ephemeris]]
+PSRB            1957+20
+RAJ             19:59:36.7621468         1  0.00006181131892971835
+DECJ           +20:48:14.89600           1  0.00128570913255993488
+F0             622.12202878404123374     1  0.00000000002809749898
+F1             -6.4968101699156749431e-15 1  7.6260018607009905757e-20
+PEPOCH         51260
+POSEPOCH       51260
+DMEPOCH        51260
+DM             29.123997272213708468     1  0.00005644573927968912
+PMRA           -17.040024613116855064    1  0.07695426370398671112
+PMDEC          -24.727440629601155754    1  0.11262609620175098590
 BINARY         ELL1
-PB             0.38196657360519661225    1  0.00000000213643338414   
-A1             0.089227042390331765365   1  0.00000040732298994754   
-PBDOT          3.6900860246209443993e-11 1  5.2557255178724327592e-13
-TASC           51260.200706741952363     1  0.00001119431690960640   
-EPS1           0.00010880456643098245148 1  0.00000840666496396327   
-EPS2           0.00011017880872190852658 1  0.00000733921385557160   
-START          54338.036962907237466       
-FINISH         56428.397817063043476       
-TZRMJD         55404.096064753292108       
-TZRFRQ         314.95999999999997954    
-TZRSITE        i   
-TRES           15.592       
-EPHVER         5                           
+PB             0.38196660517441393128    1  0.00000000143634118308
+A1             0.089227019738472404882   1  0.00000036993283659531
+PBDOT          2.9410054626902594231e-11 1  3.369870619559308697e-13
+TASC           51260.2005372410819       1  0.00000785660295236510
+EPS1           3.993763661004633939e-05  1  0.00000766497092897124
+EPS2           5.1836632726032813506e-05 1  0.00000704805653354374
+START          54338.036962885656976
+FINISH         56885.154009167665208
+TZRMJD         55614.588078623214397
+TZRFRQ         341.24900000000002365
+TZRSITE        i
+TRES           17.025
+EPHVER         5
 CLK            TT(TAI)
 MODE 1
 UNITS          TDB
+TIMEEPH        IF99
+DILATEFREQ     Y
+PLANET_SHAPIRO Y
+T2CMETHOD      IAU2000B
+NE_SW          4.000
+CORRECT_TROPOSPHERE  Y
 EPHEM          DE405
 NITS           1
-NTOA           5642
-CHI2R          90.1647 5628
+NTOA           7029
+CHI2R          95.7907 7015
