+++ scintellometry/io/fitsdef.txt
