+++ scintellometry/io/fromfile.py
@@ -8,7 +8,8 @@ msblsb_bits = np.array([-16, 15], np.int8)
 twopiby256 = 2.*np.pi / 256.
 
 NP_DTYPES = {'1bit': 'i1', '4bit': 'i1', 'nibble': 'i1',
-             'ci1': '2i1', 'ci1,ci1': '2i1,2i1', 'cu4bit': 'u1'}
+             'ci1': '2i1', 'ci1,ci1': '2i1,2i1', 'cu4bit,cu4bit': '2u1'}
+
 
 def fromfile(file, dtype, count, verbose=False):
     """Read count bytes, with type dtype which can be bits.
@@ -69,6 +70,23 @@ def fromfile(file, dtype, count, verbose=False):
         # so least significant bits go first.
         np.right_shift(split, 4, split)  # explicitly give output for speedup
         return split
+    elif dtype.startswith('cu4bit'):
+        # For a given int8 byte containing bits 76543210, with real as 
+        # 4 most significant bits and imaginary as 4 least significant bits:
+        # right_shift(byte[:,np.newaxis], shift40):  [xxxx7654, 76543210]
+#        raw = raw.flatten()
+#        raw_re = ((raw >> 4).astype(np.int8) - 8)
+#        raw_im = ((raw & 0xf).astype(np.int8) - 8)
+
+#        data = np.zeros([2*len(raw_re)], dtype=np.int8)
+#        data[::2] = raw_re
+#        data[1::2] = raw_im
+
+        split = np.right_shift(raw.reshape(-1, 1), shift40).flatten()
+        split &= 0xf
+        # now ----7654, ----3210
+        split -= 8
+        return split.astype('f4').view(dtype.replace('cu4bit', 'c8')).squeeze()
     elif dtype == 'nibble':
         # For a given int8 byte containing bits 76543210
         split = np.bitwise_and(raw[:,np.newaxis], msblsb_bits)
@@ -87,14 +105,6 @@ def fromfile(file, dtype, count, verbose=False):
         amp = np.sqrt(-2.*np.log(1.-amp))
         phase = (iph.astype(np.float32) + 8.) * twopiby256
         return amp * np.exp(1.j * np.pi * phase)
-#    elif dtype == 'cu4bit':
-#        raw_re = ((raw // 16).view(np.int8) - 8).astype(np.int32)
-#        raw_im = ((raw % 16).view(np.int8) - 8).astype(np.int32)
-#        return (raw_re + 1.0j * raw_im)
-    elif dtype == 'cu4bit':
-        raw_re = ((raw >> 4).astype(np.int32) - 8)
-        raw_im = ((raw & 0xf).astype(np.int32) - 8)
-        return (raw_re + 1.0j * raw_im)
     else:
         # this should never happen, but just in case...
         raise TypeError('data type "{}" not understood (but in NP_DTYPES!)'
