+++ scintellometry/io/multifile.py
