+++ scintellometry/trials/vlbi_b0329/observations.conf
@@ -1,0 +1,64 @@
+# we have two main section, [psrs] describing the sources
+# the rest detail observation at a particular telescope
+# and have structure
+# [telescope name]
+#     options_as_parsed_in_observations.py
+#     [[an observation]]
+#     src = psr
+#     extra_options_as_parsed_in_observations.py
+
+# Notes: actual observations must have a key consistent
+#       with astropy.time.Time
+#       all observations need a 'src' key 
+#    _____  ___   __ __  ____      __    ___  _____
+#   / ___/ /   \ |  |  ||    \    /  ]  /  _]/ ___/
+#  (   \_ |     ||  |  ||  D  )  /  /  /  [_(   \_ 
+#   \__  ||  O  ||  |  ||    /  /  /  |    _]\__  |
+#   /  \ ||     ||  :  ||    \ /   \_ |   [_ /  \ |
+#   \    ||     ||     ||  .  \\     ||     |\    |
+#    \___| \___/  \__,_||__|\_| \____||_____| \___|
+#                                                  
+
+[psrs]    
+  [[0809+74]]
+      dm = 6.116 * u.pc / u.cm**3
+      coords = '08h09m 74d00m'
+  [[1508+55]]
+      dm = 19.613 * u.pc / u.cm**3
+  [[B0329+54]]
+      dm = 26.833 * u.pc / u.cm**3
+#      dm = 500.000 * u.pc / u.cm**3
+  [[B0823+26]]
+      dm = 19.454 * u.pc / u.cm**3
+  [[J1810+1744]]
+      # dm = 39.659298 * u.pc / u.cm**3
+      dm = (39.7+0.035) * u.pc / u.cm**3
+  [[B1919+21]]
+      dm = 12.455   * u.pc / u.cm**3
+  [[B1957+20]]
+      dm = 29.11680*  1.001 * u.pc / u.cm**3
+  [[B2016+28]] 
+      dm = 14.172 * u.    pc / u.cm**3
+  [[B2111+46]]
+      dm = 141.26 * u.pc / u   .cm**3
+  [[noise]]
+      dm = 0. * u.pc / u.cm**3
+
+[arochime]
+fnbase = 'arochime'
+
+   [[setup]]
+   samplerate = 800.*u.MHz
+   fedge = 800.*u.MHz
+   fedge_at_top = True
+   # 15099494400 bytes / 3600 timestamps -> 4194304 -> 2**22 block size
+   # also the same as used by Ue-Li: ntint=8192=2**13, nblock=512=2**9
+   # reads ntint/2 nblock*2 bytes or 2**22
+   blocksize = 2**23
+
+   [[2014-08-01T11:41:38.17]]
+   src = B0329+54
+   ppol = """Polynomial([0.0, 1.39964041492])"""
+#   file_fmt = '{0}/01aug2014/b0329+54.raw.Pol_{1:1s}.dat'
+   file_fmt = '{0}/b0329_vlbi1/{1:07d}.dat'
+   last = 87
