+++ scintellometry/trials/vlbi_b0329/observations.py
@@ -1,0 +1,1 @@
+../vlbi_j1012/observations.py
