+++ scintellometry/trials/vlbi_b0329/reduction.py
@@ -1,0 +1,1 @@
+../vlbi_j1012/reduction.py
