+++ scintellometry/trials/vlbi_b1919/observations.conf
@@ -27,6 +27,7 @@
       dm = 19.613 * u.pc / u.cm**3
   [[B0329+54]]
       dm = 26.833 * u.pc / u.cm**3
+#      dm = 500.000 * u.pc / u.cm**3
   [[B0823+26]]
       dm = 19.454 * u.pc / u.cm**3
   [[J1810+1744]]
@@ -254,9 +255,11 @@ P = 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18
 fnbase = 'gmrt'
 
    [[setup]]
-   nchan = 512
-   samplerate = 200./6.*u.MHz
-   fedge = 156.*u.MHz
+   nchan = 1024
+#   samplerate = 200./6.*u.MHz
+#   fedge = 156.*u.MHz
+   samplerate = 800.*u.MHz
+   fedge = 800.*u.MHz
    fedge_at_top = True
    # check b1919+21.raw.timestamp
    # 15099494400 bytes / 3600 timestamps -> 4194304 -> 2**22 block size
@@ -314,64 +317,28 @@ fnbase = 'gmrt'
    file_fmt = '{0}/node33/28jul2013/b2111+46.raw.Pol-{1:1s}{2:1d}.dat'
    pol = 'R'  # on node33
 
+   [[2014-08-01T11:41:38.17]]
+   src = B0329+54
+   ppol = """Polynomial([0.0, 1.39964041492])"""
+   file_fmt = '{0}/01aug2014/b0329+54.raw.Pol_{1:1s}{2:1d}.dat'
+   pol = 'R'
 
-#           _____   ____       _     
-#     /\   |  __ \ / __ \     | |    
-#    /  \  | |__) | |  | | ___| |__  
-#   / /\ \ |  _  /| |  | |/ __| '_ \ 
-#  / ____ \| | \ \| |__| | (__| | | |
-# /_/    \_\_|  \_\\____/ \___|_| |_|
 
-[AROch]
-fnbase='ARO'
-   [[setup]]	
-   blocksize=2**24
+[arochime]
+fnbase = 'arochime'
 
-   [[2014-06-13T01:36:19]]
-   data_dir = /scratch/p/pen/klages/ARO/pulsar_140613_b1957p20_v0/
-   tend = 2014-06-13T01:53:48
-   src = B1957+20
-   dirdate = 140613
-   v = 0
-   file_fmt = '{0}/pulsar_{1}_{2}_v{3}/{4}.dat'
-   ppol = """Polynomial([-1252679.1986725251,                                                                                                                                                         
-                          601.39629721056895,                                                                                                                                                          
-                           -6.6664639926379228e-06,                                                                                                                                                    
-                           -3.005404797321569e-10,                                                                                                                                                     
-                            1.3404520057431192e-13,                                                                                                                                                    
-                            3.5632030706667189e-18,                                                                                                                                                    
-                           -1.0874017282180807e-21,                                                                                                                                                    
-                           -1.8089896985287676e-26,                                                                                                                                                    
-                            4.803545433801123e-30,                                                                                                                                                     
-                            1.4787240038933893e-35,                                                                                                                                                    
-                           -1.1792841185454315e-38,                                                                                                                                                    
-                            2.6298912108944255e-43])"""
+   [[setup]]
+   samplerate = 800.*u.MHz
+   fedge = 800.*u.MHz
+   fedge_at_top = True
+   # check b1919+21.raw.timestamp
+   # 15099494400 bytes / 3600 timestamps -> 4194304 -> 2**22 block size
+   # also the same as used by Ue-Li: ntint=8192=2**13, nblock=512=2**9
+   # reads ntint/2 nblock*2 bytes or 2**22
+   blocksize = 2**23
 
-   [[2014-06-13T02:07:55]]
-   data_dir = /scratch/p/pen/klages/ARO/pulsar_140613_b1957p20_v0/
-   tend = 2014-06-13T02:12:48
-   src = B1957+20
-   dirdate = 140613
-   v = 1
-   file_fmt = '{0}/pulsar_{1}_{2}_v{3}/{4}.dat'
-   ppol = None
-
-   [[2014-06-13T02:31:26]]
-   tend = 2014-06-13T03:12:40
-   data_dir = /scratch/p/pen/klages/ARO/pulsar_140613_b1957p20_v0/
-   src = B1957+20
-   dirdate = 140613
-   v = 2
-   file_fmt = '{0}/pulsar_{1}_{2}_v{3}/{4}.dat'
-   ppol = None
-
-   [[2014-06-13T03:18:57]]
-   tend = 2014-06-13T03:30:08
-   data_dir = /scratch/p/pen/klages/ARO/pulsar_140613_b1957p20_v0/
-   src = B1957+20
-   dirdate = 140613
-   v = 3
-   file_fmt = '{0}/pulsar_{1}_{2}_v{3}/{4}.dat'
-   ppol = None
-
-# LocalWords:  dirdate
+   [[2014-08-01T11:41:38.17]]
+   src = B0329+54
+   ppol = """Polynomial([0.0, 1.39964041492])"""
+   file_fmt = '{0}/01aug2014/b0329+54.raw.Pol_{1:1s}.dat'
+   pol = 'R'
